Face Detection
The objective of the program given is to detect object of interest(face) in real time and to keep tracking of the same object.This is a simple example of how to detect face in Python. You can try to use training samples of any other object of your choice to be detected by training the classifier on required objects.

 Steps:

Download Python 2.7.x version, numpy and Opencv 2.7.x version.Check if your Windows either 32 bit or 64 bit is compatible and install accordingly.
Make sure that numpy is running in your python then try to install opencv.
Put the haarcascade_eye.xml & haarcascade_frontalface_default.xml files in the same folder.